+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Add AVHRRMTA_G-NAVO-L2P-v1.0 to associations
     - Added in shared memory allocation limit in fork process
   - PODAAC-4173
-    - Add AVHRRMTB_G-NAVO-L2P-v1.0 to associations 
+    - Add AVHRRMTB_G-NAVO-L2P-v1.0 to associations
+  - Github Issue 10
+    - Handle empty granule files.
 ### Changed 
 ### Deprecated 
 ### Removed